Atualização dia 07/06

+++ api/src/controllers/LockController.js
@@ -1,5 +1,7 @@
 const Lock = require ("../models/Lock");
 const Group = require ("../models/Group");
+const LocalFisico = require ("../models/LocalFisico");
+
 
 module.exports = {
 
@@ -16,37 +18,71 @@ module.exports = {
 
     async store (request, response){
 
-        const {name, _id = null} = request.body;
+        const {name, _id = null, Localtype="group"} = request.body;
         
-        const holderGroup = await Group.findById(_id);
+        if(Localtype==="group"){
 
-        let lock = await Lock.findOne({ name });
+            const holderGroup = await Group.findById(_id);
 
-        if(lock===null){
-            
-            if(holderGroup!==null){
+            let lock = await Lock.findOne({ name });
 
-                var newHolder = holderGroup.holder;
-                newHolder.push(holderGroup._id);
+            if(lock===null){
+                
+                if(holderGroup!==null){
 
-                const NewLock = await Lock.create({
-                    name,
-                    holder: newHolder
+                    var newHolder = holderGroup.holder;
+                    newHolder.push(holderGroup._id);
+
+                    const NewLock = await Lock.create({
+                        name,
+                        holder: newHolder
+                    });
+
+                    let newContent = holderGroup.locks;
+                    newContent.push(NewLock._id);
+                    await Group.findByIdAndUpdate({ _id: holderGroup._id}, { locks: newContent}, {new: true});
+
+                    return response.json ({NewLock});
+                } else return response.status(400).json({
+                    error: true,
+                    message: 'O grupo no qual deseja criar a tranca não foi encontrado'
                 });
-
-                let newContent = holderGroup.locks;
-                newContent.push(NewLock._id);
-                await Group.findByIdAndUpdate({ _id: holderGroup._id}, { locks: newContent}, {new: true});
-
-                return response.json ({NewLock});
             } else return response.status(400).json({
                 error: true,
-                message: 'O grupo no qual deseja criar a tranca não foi encontrado'
-            });
-        } else return response.status(400).json({
-            error: true,
-            message: 'Uma tranca com esse nome já existe'
-        })
+                message: 'Uma tranca com esse nome já existe'
+            })
+        }
+
+        if(Localtype==="localFisico"){
+            
+            const holderLocal = await LocalFisico.findById(_id);
+
+            let lock = await Lock.findOne({ name });
+
+            if(lock===null){
+                
+                if(holderLocal!==null){
+
+                    const NewLock = await Lock.create({
+                        name,
+                        holder: holderLocal.holder,
+                        holderLocalFisico: holderLocal._id
+                    });
+
+                    let newContent = holderLocal.locks;
+                    newContent.push(NewLock._id);
+                    await LocalFisico.findByIdAndUpdate({ _id: holderLocal._id}, { locks: newContent}, {new: true});
+
+                    return response.json ({NewLock});
+                } else return response.status(400).json({
+                    error: true,
+                    message: 'O local físico no qual deseja criar a tranca não foi encontrado'
+                });
+            } else return response.status(400).json({
+                error: true,
+                message: 'Uma tranca com esse nome já existe'
+            })
+        }
     },
 
     async Update (request, response)
@@ -58,9 +94,16 @@ module.exports = {
 
     async destroy (request, response)
     {
-        const {_id} = request.headers;
-        const lock = await Lock.findByIdAndDelete(_id);
-        var newContentGroup = await Group.findOneAndUpdate({locks: {$in: [_id]}}, {$pullAll: {locks: [_id]}}, {new: true});
-        return response.json ({lock, newContentGroup});
+        const {_id, Localtype="group"} = request.headers;
+        if(Localtype==="group"){
+            const lock = await Lock.findByIdAndDelete(_id);
+            var newContentGroup = await Group.findOneAndUpdate({locks: {$in: [_id]}}, {$pullAll: {locks: [_id]}}, {new: true});
+            return response.json ({lock, newContentGroup});
+        } 
+        if(Localtype==="localFisico"){
+            const lock = await Lock.findByIdAndDelete(_id);
+            var newContentLocal = await LocalFisico.findOneAndUpdate({locks: {$in: [_id]}}, {$pullAll: {locks: [_id]}}, {new: true});
+            return response.json ({lock, newContentLocal});
+        }
     }
 };
